add sample survey update component
