+++ packages/user-ui/src/pages/project/components/mod.rs
@@ -1,6 +1,5 @@
 pub mod comments;
 pub mod multiple_objective;
-pub mod not_complete_survey_modal;
 pub mod sample;
 pub mod single_objective;
 pub mod subjective;
