+++ packages/user-ui/src/pages/project/i18n.rs
@@ -15,6 +15,10 @@ translate! {
         ko: "저장하고 나가기",
         en: "Save and Exit"
     }
+    remove_modal_title: {
+        ko: "내가 작성한 답변을 삭제합니다.",
+        en: "I will delete the answer I wrote."
+    }
 }
 
 translate! {
@@ -23,6 +27,18 @@ translate! {
     sample_survey: {
         ko: "표본 조사",
         en: "Sample Survey"
+    }
+    written_by_me: {
+        ko: "내가 작성한 답변",
+        en: "Answer written by me"
+    }
+    responses: {
+        ko: "질문별 응답",
+        en: "Responses to each question"
+    }
+    see_detail: {
+        ko: "자세히 보기",
+        en: "See Detail"
     }
 }
 
@@ -36,5 +52,22 @@ translate! {
     input_hint: {
         ko: "내용을 입력해주세요",
         en: "Please Enter Details"
+    }
+}
+
+translate! {
+    MySurveyTranslate;
+
+    title: {
+        ko: "내가 작성한 답변",
+        en: "Answer written by me"
+    }
+    update: {
+        ko: "수정하기",
+        en: "Update"
+    }
+    remove: {
+        ko: "삭제하기",
+        en: "Remove"
     }
 }
